requirements/base.txt
CairoSVG==2.7.1
Markdown==3.7
Pillow==11.0.0
PyYAML==6.0.2
Whoosh==2.7.4
bleach==6.2.0
boto3==1.35.85
celery==5.4.0
dateparser==1.2.0
django-activity-stream==2.0.0
django-auth-ldap==5.1.0
django-celery-beat==2.7.0
django-cors-headers==4.6.0
django-otp==1.3.0
django-two-factor-auth==1.17.0
django-formtools==2.3
django-mathfilters==1.0.0
django-model-utils==5.0.0
django-mptt==0.16.0
django-qsstats-magic==1.1.0
django-solo==2.4.0
django-storages==1.14.6
django-stronghold==0.4.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djangorestframework-recursive==0.1.2
drf-yasg==1.21.8
elasticsearch==7.17.12
elasticsearch-dsl==7.4.1
extract-msg==0.52.0
flex==6.14.1
furl==2.1.4
fusepy==3.0.1
gevent==24.11.1
google-cloud-storage==2.19.0
graphviz==0.20.3
greenlet==3.1.1
gunicorn==23.0.0
importlib-metadata==8.5.0
jsonschema==4.23.0
mozilla-django-oidc==4.0.1
nh3==0.2.18
node-semver==0.9.0
ollama==0.4.5
pycountry==24.6.1
pycryptodome==3.21.0
pyotp==2.9.0
pypdf==5.1.0
python-dateutil==2.9.0.post0
python_gnupg==0.5.4
pytz==2024.2
qrcode==8.0
requests==2.32.4
sentry-sdk==2.35.0
sh==2.1.0
swagger-spec-validator==3.0.4
whitenoise==6.8.2
